requirements.txt
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
pycryptodome==3.9.9
pyOpenSSL==20.0.1
six==1.15.0
Werkzeug==1.0.1
